Phase3: Fix Threading PCH, ARC flags, macOS headers, SDL/Vulkan guards
- Add common/Threading.h to PrecompiledHeader.h
- Add -fobjc-arc flag for iOS .mm files in ios/platform/CMakeLists.txt
- Wrap IOKit includes and InhibitScreensaver with #if !TARGET_OS_IOS in DarwinMisc.cpp
- Move SDL audio sources to if(NOT IOS) block in pcsx2/CMakeLists.txt
- Fix WebPDecodeRGBAInto stride argument cast in Image.cpp
- Wrap DebugTools/ccc linking with if(NOT IOS) in pcsx2/CMakeLists.txt

+++ common/Image.cpp
@@ -700,7 +700,7 @@ bool WebPBufferLoader(RGBA8Image* image, const void* buffer, size_t buffer_size)
 	std::vector<u32> pixels;
 	pixels.resize(static_cast<u32>(width) * static_cast<u32>(height));
 	if (!WebPDecodeRGBAInto(static_cast<const u8*>(buffer), buffer_size, reinterpret_cast<u8*>(pixels.data()),
-			sizeof(u32) * pixels.size(), sizeof(u32) * static_cast<u32>(width)))
+			sizeof(u32) * pixels.size(), static_cast<int>(sizeof(u32) * static_cast<u32>(width))))
 	{
 		Console.Error("WebPDecodeRGBAInto() failed");
 		return false;
